Add workflow for end-to-end test (using Docker) and native builds (#58)
* Add workflow that builds Docker image and runs it, builds TTL for Malta and checks its validity, builds QLever index and counts the number of geometries

* Add native build workflow on Ubuntu 22.04 and latest, both on clang and gcc

* Use const static inline vectors, not arrays, in ttl/Constants.h to avoid issues with implementation-dependend initialization

* Use cmake's CheckIPOSupported and INTERPROCEDURAL_OPTIMIZATION to set -flto compiler-independently

---------

+++ src/osm/GeometryHandler.cpp
@@ -60,8 +60,8 @@ GeometryHandler<W>::GeometryHandler(const osm2rdf::config::Config& config,
     : _config(config),
       _writer(writer),
       _sweeper(
-          {omp_get_max_threads(),
-           omp_get_max_threads(),
+          {static_cast<size_t>(omp_get_max_threads()),
+           static_cast<size_t>(omp_get_max_threads()),
            "",
            osm2rdf::ttl::constants::IRI__OPENGIS_INTERSECTS,
            osm2rdf::ttl::constants::IRI__OPENGIS_CONTAINS,
@@ -216,7 +216,7 @@ template <typename W>
     const ::util::geo::DPoint& loc) {
   auto point = ::util::geo::latLngToWebMerc(
       ::util::geo::DPoint(loc.getX(), loc.getY()));  // locs are lon/lat
-  return ::util::geo::I32Point{point.getX() * PREC, point.getY() * PREC};
+  return ::util::geo::I32Point{static_cast<int>(point.getX() * PREC), static_cast<int>(point.getY() * PREC)};
 }
 
 // ____________________________________________________________________________
